refactor(mux): move active connection state into module
